rename style.js to style_loginsignup, modified style
