+++ screen/welcome_business.js
@@ -12,7 +12,7 @@ import {
 
 } from 'react-native';
 
-import { colours, globalStyles } from '../components/style';
+import { colours, globalStyles } from '../components/style_loginsignup';
 //keyboardavoidingwrapper
 import KeyboardAvoidingWrapper from '../components/KeyboardAvoidingWrapper';
